+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Upgrade to pulumi-terraform-bridge v1.5.2
+* Upgrade to v0.2.3 of the Kafka Terraform Provider
 
 ---